Update xlsxworkbook.cpp
any time we no need it , the selected mark will be set in each sheet by selected="1" or "0"  
if use activeTab in workbook.xml, when the  sheet 0 is selected and write data to sheet 0 or sheet 1 will lead sheet 0 and sheet 1 are selected. 
other side, when multiple sheets are selected , the activeTab will be empty in workbook.xml.
so need this.